Use decimal datatype for money

+++ lib/auctions/api/dto/bid.rb
@@ -7,7 +7,7 @@ module Auctions
         attribute :id, Types::UUID
         attribute :bidder_id, Types::UUID
         attribute :auction_id, Types::UUID
-        attribute :amount, Types::Float
+        attribute :amount, Types::Decimal
       end
     end
   end
